cleanup of telegraf plan
This commit cleans up some of the issues with the proposed telegraf
plan.

First off, we're building telegraf from source because while the
upstream project provides prebuilt binaries, they are dynamically
linked against libc, which is not in a location that Habitat knows
about, and it's easy enough to build, rather than using `patchelf` to
modify the binary.

A README.md is added. Telegraf itself is highly customizable, and any
default configuration is hardly sufficient, but parameterizing the
world is not maintainable. The README.md describes rudimentary steps
to create a custom package that depends on this one in order to
provide a custom configuration.

A default configuration generated with `telegraf config` is added so
that the service starts up by default.

The plan itself is updated to have the latest version of telegraf,
specify the license with the proper SPDX designation, and point to the
right binary and default configuration file for the service. The
callback functions are rewritten a bit to use `go get` to download,
and use `git checkout` to "unpack", and `install` to install the
binary.

Signed-off-by: Joshua Timberman <joshua@chef.io>